Unified the pthread and verona benchmarks.
